feat: :necktie: Added controller by books and restrict delete in others models

+++ app/Http/Controllers/API/V1/Library/AuthorController.php
@@ -54,6 +54,13 @@ class AuthorController extends Controller
 
     public function destroy(Author $author): JsonResponse
     {
+        if ($author->books) {
+            return ApiResponseService::error(
+                'Author has one or more books related',
+                Response::HTTP_CONFLICT,
+            );
+        }
+
         $author->delete();
 
         return ApiResponseService::success(
